Modify code warning
Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I8Z8V0

+++ ecmascript/js_regexp_iterator.cpp
@@ -88,8 +88,8 @@ JSTaggedValue JSRegExpIterator::Next(EcmaRuntimeCallInfo *argv)
                     lastIndexString).GetValue());
                 JSTaggedNumber thisIndex = JSTaggedValue::ToLength(thread, getLastIndex);
                 RETURN_EXCEPTION_IF_ABRUPT_COMPLETION(thread);
-                uint32_t nextIndex = BuiltinsRegExp::AdvanceStringIndex(inputStr, thisIndex.ToUint32(),
-                                                                        fullUnicode);
+                uint32_t nextIndex = static_cast<uint32_t>(
+                    BuiltinsRegExp::AdvanceStringIndex(inputStr, thisIndex.ToUint32(), fullUnicode));
                 ObjectFastOperator::FastSetPropertyByValue(thread, regexHandle.GetTaggedValue(),
                                                            lastIndexString.GetTaggedValue(),
                                                            JSTaggedValue(nextIndex));
